Trying to get delete to work. Also merging.
Hoorah!

+++ basketballGAME/basketballUI/TeamManagement.xaml.cs
@@ -151,16 +151,43 @@ public partial class TeamManagement : ContentPage
     async private void DeleteTeamButton_Clicked(object sender, EventArgs e)
     {
         string API_URL = "http://localhost:5121/api/Teams";
+        string API_URL_TEAMPLAYERS = "http://localhost:5121/api/TeamPlayers/";
 
         using (HttpClient client = new HttpClient())
         {
             try
             {
-                HttpResponseMessage response3 = await client.DeleteAsync($"{API_URL}/{TeamIDDelete.Text}");
+                HttpResponseMessage response1 = await client.GetAsync(API_URL_TEAMPLAYERS);
+                List<TeamPlayer> teamPlayers;
 
-                if (response3.IsSuccessStatusCode)
+                if (response1.IsSuccessStatusCode)
                 {
-                    DeleteTeamButton.Text = "Deleted";
+                    string json1 = await response1.Content.ReadAsStringAsync();
+                    teamPlayers = JsonConvert.DeserializeObject<List<TeamPlayer>>(json1);
+
+                    foreach (TeamPlayer tp in teamPlayers)
+                    {
+                        if (tp.TeamId == int.Parse(TeamIDDelete.Text)) 
+                        {
+                            HttpResponseMessage response2 = await client.DeleteAsync($"{API_URL_TEAMPLAYERS}/{tp.Id}");
+                            if (response2.IsSuccessStatusCode) 
+                            {
+                                DeleteTeamButton.Text = "SUCCESS!";
+                                return;
+                            }
+                            else
+                                DeleteTeamButton.Text = "Not deleted.\n" + response2;
+                        }
+                    }
+
+                    HttpResponseMessage response3 = await client.DeleteAsync($"{API_URL}/{TeamIDDelete.Text}");
+
+                    if (response3.IsSuccessStatusCode)
+                    {
+                        DeleteTeamButton.Text = "Deleted";
+                    }
+                    else
+                        DeleteTeamButton.Text = "Not deleted.\n" + response3;
                 }
             }
             catch (Exception ex)
